==> notebooks/05_population_study.ipynb <==
import pandas as pd

#Import the consolidated data set
file_path = "data\\3_intermediate\\int_texas_border_report.csv"

border_report = pd.read_csv(
    file_path, 
    parse_dates=['nibrs_start_date', 'report_year'], 
    date_format='%Y-%m-%d')

border_report.head()

# Filter the border_report where population is 0 and for the most recent report year
zero_population = border_report[(border_report['population'] == 0) & (border_report['report_year'] == '2023')]

# Count unique agency names per agency type for the most recent year
zero_population_agencies = zero_population.groupby('agency_type')['agency_name'].nunique()

#Convert zero_population_agencies to a dataframe
zero_population_agencies_df = zero_population_agencies.to_frame().reset_index()

#Calculate the percentage of agencies in zero_population_agencies_df and round to two decimal places
zero_population_agencies_df['percentage'] = round((zero_population_agencies_df['agency_name'] / zero_population_agencies_df['agency_name'].sum()) * 100, 2)

#Sort zero_population_agencies_df by percentage in descending order
zero_population_agencies_df = zero_population_agencies_df.sort_values(by='percentage', ascending=False)

#Print total number of agencies with zero population
print(f"Total number of agencies with zero population: {zero_population_agencies_df['agency_name'].sum()}")
print()
print(zero_population_agencies_df.to_markdown(index=False))


# Drop rows where the population is zero
populated_agencies = border_report[border_report['population'] > 0]

# Group by county and report_year, the calculation the population for the county
county_population = populated_agencies.groupby(['county', 'report_year'])['population'].sum().reset_index()

# Pivot for easier analysis
population_pivot = county_population.pivot(index='county', columns='report_year', values='population')

# Calculate the overall percent change and numerical change
first_year, last_year = population_pivot.columns[0], population_pivot.columns[-1]
overall_percent_change = (population_pivot[last_year] - population_pivot[first_year]) / population_pivot[first_year] * 100
numerical_change = population_pivot[last_year] - population_pivot[first_year]

# Adjustments as per requirements
overall_percent_change_rounded = overall_percent_change.round(2)
numerical_change_int = numerical_change.astype(int)
most_recent_year_population_int = population_pivot[last_year].astype(int)

# Assemble the final table
final_table = pd.DataFrame({
    'County': population_pivot.index,
    'Latest Population': most_recent_year_population_int.values,
    'Numerical Change': numerical_change_int.values,
    'Percent Change': overall_percent_change_rounded.values
}).reset_index(drop=True)

#Sort final_table by Numerical Change in descending order
final_table = final_table.sort_values(by='Numerical Change', ascending=False)

#Convert final_table to .csv. #Sort by percent change
final_table.to_csv('data\\4_fact\\average_percent_change_in_population.csv', index=False)

#Calculate the average percentage change in population
average_percent_change = overall_percent_change.mean()
print(f"Average percentage change in population: {average_percent_change:.2f}%")

print(final_table.sort_values(by='Percent Change', ascending=False).head(n=3).to_markdown(index=False))

#Calculate the total population in 2023
total_population_2023 = population_pivot[last_year].sum()
print(f"Total population in 2023: {total_population_2023:,}")

#Return the first 3 rows of final_table and turn off the index
final_table.head(n=3)
